Raise checked error for XML parsing error.
Previously runtime error is raised if XMLStreamException is raised from the StAX
parsing library, which is not very useful to determine what is the real cause.
Added a check of the location info of the XMLStreamException instance, if the
location info is there, the exception is actually caused by malformed file and
a XmlParseException will be raised to make it more clear to the caller.

Change-Id: Icaf24493c6a0f4c4835a02bc9b93ce2df273fe35