Use newer version of proto java bindings, and allow more proto versions.
Experimentally, I've determined:
Old versions of protoc (i.e. v2.4) work with new versions of the Java
bindings. However, new versions of protoc don't work with old versions
of the Java bindings. So I've updated to the version of the Java
bindings according to the newest released version of protoc (v2.6.0),
and updated the toolchains.xml file to allow for both new and old
versions of protoc.

Change-Id: I924b93afcd129211bf8818acbc04edc6436511f0